Move `WithLeaderSchedule` to ouroboros-consensus-mock.
`WithLeaderSchedule` is morally just part of the mock protocol as it overrides
most of the functionalities inside `ConsensusProtocol` and it is only used in
the mock implementation.

`LeaderSchedule` is left behind so that we don't need to make
`ouroboros-consensus-test` depend on `ouroboros-consensus-mock`.
