+++ ouroboros-consensus-mock/ouroboros-consensus-mock.cabal
@@ -44,6 +44,7 @@ library
                        Ouroboros.Consensus.Mock.Node.Praos
                        Ouroboros.Consensus.Mock.Node.PraosRule
                        Ouroboros.Consensus.Mock.Node.Serialisation
+                       Ouroboros.Consensus.Mock.Protocol.LeaderSchedule
                        Ouroboros.Consensus.Mock.Protocol.Praos
 
   build-depends:       base              >=4.9   && <4.15
