+++ ouroboros-consensus/src/Ouroboros/Consensus/Protocol/LeaderSchedule.hs
@@ -6,9 +6,7 @@
 {-# LANGUAGE TypeFamilies       #-}
 
 module Ouroboros.Consensus.Protocol.LeaderSchedule (
-    ConsensusConfig (..)
-  , LeaderSchedule (..)
-  , WithLeaderSchedule
+    LeaderSchedule (..)
   , leaderScheduleFor
   ) where
 
@@ -20,8 +18,6 @@ import           NoThunks.Class (NoThunks)
 
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.NodeId (CoreNodeId (..), fromCoreNodeId)
-import           Ouroboros.Consensus.Protocol.Abstract
-import           Ouroboros.Consensus.Ticked
 import           Ouroboros.Consensus.Util.Condense (Condense (..))
 
 {-------------------------------------------------------------------------------
@@ -57,43 +53,3 @@ instance Condense LeaderSchedule where
     condense (LeaderSchedule m) = condense
                                 $ map (\(s, ls) -> (s, map fromCoreNodeId ls))
                                 $ Map.toList m
-
-{-------------------------------------------------------------------------------
-  ConsensusProtocol instance
--------------------------------------------------------------------------------}
-
--- | Extension of protocol @p@ by a static leader schedule.
-data WithLeaderSchedule p
-
-data instance ConsensusConfig (WithLeaderSchedule p) = WLSConfig {
-      wlsConfigSchedule :: !LeaderSchedule
-    , wlsConfigP        :: !(ConsensusConfig p)
-    , wlsConfigNodeId   :: !CoreNodeId
-    }
-  deriving (Generic)
-
-instance ConsensusProtocol p => ConsensusProtocol (WithLeaderSchedule p) where
-  type SelectView    (WithLeaderSchedule p) = SelectView p
-
-  type ChainDepState (WithLeaderSchedule p) = ()
-  type LedgerView    (WithLeaderSchedule p) = ()
-  type ValidationErr (WithLeaderSchedule p) = ()
-  type IsLeader      (WithLeaderSchedule p) = ()
-  type ValidateView  (WithLeaderSchedule p) = ()
-  type CanBeLeader   (WithLeaderSchedule p) = ()
-
-  protocolSecurityParam = protocolSecurityParam . wlsConfigP
-
-  checkIsLeader WLSConfig{..} () slot _ =
-    case Map.lookup slot $ getLeaderSchedule wlsConfigSchedule of
-        Nothing -> error $ "WithLeaderSchedule: missing slot " ++ show slot
-        Just nids
-            | wlsConfigNodeId `elem` nids -> Just ()
-            | otherwise                   -> Nothing
-
-  tickChainDepState     _ _ _ _ = TickedTrivial
-  updateChainDepState   _ _ _ _ = return ()
-  reupdateChainDepState _ _ _ _ = ()
-
-instance ConsensusProtocol p
-      => NoThunks (ConsensusConfig (WithLeaderSchedule p))
